Let devs set field options from dashboard classes
This is a prerequisite for extending the belongs_to and has_many
fields to take options such as alternate table names, etc.

Other potential uses could be to specify the number of decimal places to
show a floating-point number to, or whether or not to truncate a string.

- Update generators to new attribute_types syntax

Cards that depend on this commit:
https://trello.com/c/kypYTxFt
https://trello.com/c/GDw9Luy3