Add comments to page views and field partials
https://trello.com/c/2Pan7Jmp

Problem:

When a developer generates a template in order to override it,
it is not always clear which local and instance variables are available
for them to use.

In order to get a good idea of what's going on,
they would need to open up Administrate's source code,
take a look at the controller,
and see which variables are being assigned for the view to use.

Solution:

Add comments at the top of each view
detailing which variables are in scope,
with links to further documentation about the variables.