Add basic validations for Product slugs
Problem:

On the demo site, it's rather simple for users to craft product names
that don't translate into valid slugs.

When Rails tries to render links to these products, it throws a 500
error because it cannot generate a correct URL.

Solution:

Add validators to ensure that all products have a valid and unique slug.