Extract an `administrate:manifest` generator
Problem:

In established applications with many models,
there is no easy way to generate Administrate dashboards
for some but not all of the models.

Solution:

Add a generator for creating only an Administrate dashboard manifest.
Users can edit the manifest file to add or remove models,
and then run `rails g administrate:install`
to generate only the dashboards listed in the manifest.

If the user runs `rails g administrate:install`
without first generating a manifest file,
Administrate will automatically create a manifest
and use it to guide the installation.

Small Changes:

- Stub out `Rails.root` in generators during tests.
  Using `Rails.root` lets us more accurately generate files
  when the Rails app is not at the root of the repository,
  such as in the example app.