Use default styling on select elements
There isn't a standardized way of customizing the look and feel of select
elements. It's always a bit of a hack. We are also still using default checkmark
and radio buttons, so why only customize some of the default form fields?

For now, let's go back to using default select and we can revisit customizing
the look and feel in the future, after some CSS clean up.