Localise Dates (#570)
* Reorder I18n backends in `with_translations` helper

The `with_translations` helper allows translations to be setup for a
method under test. A recent change (#772) added a backend chain to
ensure this helper won't break other spec by replacing all translations
and causing "missing_translations" failures.

This changes the order of the I18n backends so that specs will first use
the translations provided to the `with_translations` block and then
fallback to the default translations. Otherwise, there would be no way
to override the default translations in test cases.

* Adds format option to DateTime field

Updates DateTime field to accept 'format' option which determines
the display format of the field. The option can be
* A key for the localized date format in the corresponding yaml config
file
* A strftime format string