Test against Rails 6.0.0rc1
To get ready for Rails 6.0.0 we can start testing against 6.0.0.rc1

Rails 6.0.0 has some breaking changes that don't allow it to work with
older versions of rspec-rails, so we will need to use rspec-rails 4.0.
rspec-rails 4.0 is not compatible with older versions of rails.

I also removed the pessimistic version constraints for the Rails
libraries in the gemspec. These constraints make it harder for people to
upgrade, and it is something Rails itself is moving away from.