Accessible from templates; avoids having to create new field types (#1513)
Over at https://github.com/thoughtbot/administrate/issues/612, a fair point was
made: if field `options` were public, they could be accessed from templates.
This would enable users to customize fields without having to create new field
types, which I find a bit more cumbersome.

An argument against this change would be that it increases the contact surface
of field types, which can be an issue later if we change the interface of
fields.

An argument for this change would be that the interface of fields is unlikely to
change radically at this point. If it did, probably subclasses would have to
change anyway too, possibly forcing more significant changes than that of how to
access the options.