More natural tabbing across rows in table (#1557)
When the user presses tab, the browser is expected to move to the next link.
Since all cells in the table are links, the user has to press tab many times
(once per column) in order to move to the next row:

![Tab stopping at each single cell](https://user-images.githubusercontent.com/36066/74949069-215c2000-53f5-11ea-9c38-bf605161ad26.gif)

Since each row works as a link, so much tabbing should not be necessary. This
change makes it so that each row is tabbable, but not the cells.

There are two exceptions: action links, and data that should actually be
represented as a link (eg: instances of `Field::Url` or `Field::Email`). These
are not affected by this change, and are still accessible with an additional tab
press for each cell. The row will be highlighted first, then each data link:

![Tab stopping at whole rows, and specific link cells](https://user-images.githubusercontent.com/36066/74950140-98de7f00-53f6-11ea-9752-521369b850e9.gif)