Provide a single, unified source of documentation (#1811)
Fixes https://github.com/thoughtbot/administrate/issues/1806

The "Getting started" section in the README is currently a duplicate of the
one in the docs. They also provide different level of information, which can
confuse readers who have read one but have questions solved by the other.

This MR removes this section from the README, and adds missing information to
the one in the docs. This should help us maintain a single, unified, and
reliable source of documentation.