Add guides for Fields::Url (#1937)
During work on a project using administrate during the past year I
started implementing a URL field. While grepping the codebase for
another symbol I noticed the included Fields::Url field and realized I
didn't need to implement my own (mainly for the show/index templating).

This change adds documentation for the field so that it is more
discoverable to new users of administrate.